rename Cyto to PortalMap and move some file around
