+++ packages/frontend/src/common/data/constants.ts
@@ -1,6 +1,8 @@
-import { ZoneLight } from '../../components/ZoneSearch/zoneSearchUtils'
+import { ZoneLight } from '../../ZoneSearch/zoneSearchUtils'
 import { PortalSize } from '../../types'
 
 export const DEFAULT_PORTAL_SIZE: PortalSize = 7
 
 export const DEFAULT_ZONE: ZoneLight = { name: '', value: '' }
+
+export const BAD_PASS: string = '🙅‍♀️bad password🤦‍♂️'
