+++ packages/frontend/src/common/hooks/useGetPortals.ts
@@ -3,10 +3,10 @@ import { useCallback } from 'react'
 import { useDispatch, useSelector } from 'react-redux'
 
 import { RootState } from '../../reducers'
-import { BAD_PASS } from '../../reducers/configReducer'
 import { ErrorActionTypes } from '../../reducers/errorReducer'
 import { PortalMapActionTypes } from '../../reducers/portalMapReducer'
-import { Portal } from '../../types'
+import { BAD_PASS } from '../data/constants'
+import { Portal } from '../types'
 import fetchPortals from '../utils/fetchPortals'
 import useConfigSelector from './useConfigSelector'
 import useZoneListSelector from './useZoneListSelector'
